+++ xcat-inventory/xcclient/inventory/utils.py
@@ -86,7 +86,7 @@ def Util_setdictval(mydict,keystr,value):
 
 def loadfile(filename):
     if not os.path.exists(filename):
-        raise FileNotExistException("Error: File %s does not exist, please check..." % filename)
+        raise FileNotExistException("Error: File '%s' does not exist, please check..." % filename)
 
     contents={}
     fmt = 'json'
@@ -118,3 +118,17 @@ def initglobal():
         xcat_version=""
     globalvars.xcat_version=out.strip()
     globalvars.xcat_verno=globalvars.xcat_version.split(' ')[1]
+
+# if "key" of d1 or "key" of d1[key] not in d2, delete it
+def filter_dict_keys(d1, d2):
+    tmp_d1 = d1
+    for key in tmp_d1.keys():
+        if key not in d2:
+            del tmp_d1[key]
+            continue
+        if type(tmp_d1[key]) != dict:
+            continue
+        for subkey in tmp_d1[key].keys():
+            if subkey not in d2[key]:
+                del tmp_d1[key][subkey]
+    return tmp_d1
